+++ airbyte-cdk/python/airbyte_cdk/sources/lcc/stream_slicers/single_slice.py
@@ -4,7 +4,7 @@
 from typing import Any, Iterable, Mapping
 
 from airbyte_cdk.models import SyncMode
-from airbyte_cdk.sources.lcc.stream_slicers.StreamSlicer import StreamSlicer
+from airbyte_cdk.sources.lcc.stream_slicers.stream_slicer import StreamSlicer
 
 
 class SingleSlice(StreamSlicer):
