+++ airbyte-cdk/python/airbyte_cdk/sources/lcc/stream_slicers/stream_slicer.py
